BOOST-738 rename files to comply with Booster expected namings
