+++ packages/framework-integration-tests/src/commands/ChangeCartItem.ts
@@ -1,6 +1,6 @@
 import { Command } from '@boostercloud/framework-core'
 import { Register, UUID } from '@boostercloud/framework-types'
-import { ChangedCartItem } from '../events/changed-cart-item'
+import { CartItemChanged } from '../events/CartItemChanged'
 
 @Command({
   authorize: 'all',
@@ -9,6 +9,6 @@ export class ChangeCartItem {
   public constructor(readonly cartId: UUID, readonly productId: UUID, readonly quantity: number) {}
 
   public async handle(register: Register): Promise<void> {
-    register.events(new ChangedCartItem(this.cartId, this.productId, this.quantity))
+    register.events(new CartItemChanged(this.cartId, this.productId, this.quantity))
   }
 }
