+++ packages/framework-integration-tests/src/commands/ConfirmPayment.ts
@@ -1,6 +1,6 @@
 import { Command } from '@boostercloud/framework-core'
 import { Register, UUID } from '@boostercloud/framework-types'
-import { CartPaid } from '../events/cart-paid'
+import { CartPaid } from '../events/CartPaid'
 
 /**
  * Most payment platforms like PayPal have callbacks to confirm payment.
