+++ packages/framework-integration-tests/src/commands/UpdateShippingAddress.ts
@@ -1,7 +1,7 @@
 import { Command } from '@boostercloud/framework-core'
 import { Address } from '../common/address'
 import { UUID, Register } from '@boostercloud/framework-types'
-import { UpdatedShippingAddress } from '../events/updated-shipping-address'
+import { ShippingAddressUpdated } from '../events/ShippingAddressUpdated'
 
 @Command({
   authorize: 'all',
@@ -10,6 +10,6 @@ export class UpdateShippingAddress {
   public constructor(readonly cartId: UUID, readonly address: Address) {}
 
   public async handle(register: Register): Promise<void> {
-    register.events(new UpdatedShippingAddress(this.cartId, this.address))
+    register.events(new ShippingAddressUpdated(this.cartId, this.address))
   }
 }
