+++ packages/framework-integration-tests/src/entities/Product.ts
@@ -1,12 +1,12 @@
 import { Entity, Reduces } from '@boostercloud/framework-core'
 import { SKU } from '../common/sku'
-import { ProductCreated } from '../events/product-created'
-import { ProductUpdated } from '../events/product-updated'
-import { ProductDeleted } from '../events/product-deleted'
+import { ProductCreated } from '../events/ProductCreated'
+import { ProductUpdated } from '../events/ProductUpdated'
+import { ProductDeleted } from '../events/ProductDeleted'
 import { Money } from '../common/money'
 import { UUID } from '@boostercloud/framework-types'
 import { Picture } from '../common/picture'
-import { ProductAvailabilityChanged } from '../events/product-availability-changed'
+import { ProductAvailabilityChanged } from '../events/ProductAvailabilityChanged'
 
 /**
  * A product is the representation of a sellable unit in our sample store
