+++ packages/framework-integration-tests/src/events/CartItemChanged.ts
@@ -1,10 +1,9 @@
 import { Event } from '@boostercloud/framework-core'
 import { UUID } from '@boostercloud/framework-types'
-import { Address } from '../common/address'
 
 @Event
-export class UpdatedShippingAddress {
-  public constructor(readonly cartId: UUID, readonly address: Address) {}
+export class CartItemChanged {
+  public constructor(readonly cartId: UUID, readonly productId: UUID, readonly quantity: number) {}
 
   public entityID(): UUID {
     return this.cartId
