+++ packages/framework-integration-tests/src/events/ProductAvailabilityChanged.ts
