+++ packages/framework-integration-tests/src/events/ProductCreated.ts
