+++ packages/framework-integration-tests/src/events/ProductDeleted.ts
