+++ packages/framework-integration-tests/src/events/ProductUpdated.ts
