+++ packages/framework-integration-tests/src/events/ShippingAddressUpdated.ts
@@ -1,9 +1,10 @@
 import { Event } from '@boostercloud/framework-core'
 import { UUID } from '@boostercloud/framework-types'
+import { Address } from '../common/address'
 
 @Event
-export class ChangedCartItem {
-  public constructor(readonly cartId: UUID, readonly productId: UUID, readonly quantity: number) {}
+export class ShippingAddressUpdated {
+  public constructor(readonly cartId: UUID, readonly address: Address) {}
 
   public entityID(): UUID {
     return this.cartId
