+++ packages/framework-integration-tests/src/events/StockMoved.ts
