+++ packages/framework-integration-tests/src/read-models/product-read-model.ts
@@ -1,7 +1,7 @@
 import { ReadModel, Projects } from '@boostercloud/framework-core'
 import { User } from '../roles'
 import { UUID } from '@boostercloud/framework-types'
-import { Product } from '../entities/product'
+import { Product } from '../entities/Product'
 import { SKU } from '../common/sku'
 import { Money } from '../common/money'
